Double quotes are not properly escaped (#1)
Bazel passes to CFLAGS and CXXFLAGS the redacted values of `__DATE__`,
`__TIMESTAMP__`, and `__TIME__`.

Just a single backslash is not enough. It gets passed into the code
as redacted (without any quotes) and not "redacted". The compiler
complains about the symbol "redacted" not being found for packages
that actually use on of those fields.